requirements.txt
sqlalchemy
fastapi
uvicorn
redis
fastapi-cache2